+++ src/lib/terra-components.module.ts
@@ -24,6 +24,7 @@ import { MatDatepickerModule } from '@angular/material/datepicker';
 import { A11yModule } from '@angular/cdk/a11y';
 import { MatInputModule } from '@angular/material/input';
 import { MatPseudoCheckboxModule } from '@angular/material/core';
+import { MatTreeModule } from '@angular/material/tree';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatExpansionModule } from '@angular/material/expansion';
 
@@ -52,6 +53,7 @@ import { MatExpansionModule } from '@angular/material/expansion';
         A11yModule,
         MatSelectModule,
         MatPseudoCheckboxModule,
+        MatTreeModule,
         MatToolbarModule,
         MatExpansionModule
     ],